requirements.txt
asgiref==3.8.1
blinker==1.8.2
channels==4.1.0
click==8.1.7
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.1
django-filter==24.3
django-plotly-dash==2.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dpd-components==0.1.0
drf-yasg==1.21.7
Flask==3.0.3
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
packaging==24.1
pandas==2.2.2
plotly==5.24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
Werkzeug==3.0.4
